+++ EcoClean/ViewModel/ColetaTiporesiduoViewModel.cs
@@ -5,10 +5,10 @@ namespace EcoClean.ViewModel
 {
     public class ColetaTiporesiduoViewModel
     {
-        public int ColetaId { get; set; }
+        public long ColetaId { get; set; }
         public ColetaModel Coleta { get; set; }
 
-        public int TipoResiduoId { get; set; }
+        public long TipoResiduoId { get; set; }
         public TipoResiduoModel TipoResiduo { get; set; }
     }
 }
