+++ EcoClean/ViewModel/RotaEnderecoViewModel.cs
@@ -3,10 +3,10 @@ namespace EcoClean.ViewModel
 {
     public class RotaEnderecoViewModel
     {
-        public int EnderecoId { get; set; }
+        public long EnderecoId { get; set; }
         public EnderecoModel Endereco { get; set; }
 
-        public int RotaId { get; set; }
+        public long RotaId { get; set; }
         public RotaModel Rota { get; set; }
     }
 }
